requirements.txt
streamlit
pandas
pymongo
pydantic
pillow
pytesseract
pdfplumber
